+++ PSFramework/functions/runspace/Get-PSFDynamicContentObject.ps1
@@ -28,24 +28,16 @@
 	[OutputType([PSFramework.Utility.DynamicContentObject])]
 	[CmdletBinding(HelpUri = 'https://psframework.org/documentation/commands/PSFramework/Get-PSFDynamicContentObject')]
 	Param (
-		[Parameter(Mandatory = $true)]
+		[Parameter(Mandatory = $true, ValueFromPipeline = $true, ValueFromPipelineByPropertyName = $true)]
 		[string[]]
 		$Name
 	)
 	
-	begin
-	{
-		
-	}
 	process
 	{
 		foreach ($item in $Name)
 		{
 			[PSFramework.Utility.DynamicContentObject]::Get($Name)
 		}
-	}
-	end
-	{
-	
 	}
 }