+++ PSFramework/functions/serialization/Register-PSFTypeSerializationData.ps1
@@ -52,8 +52,4 @@
 			Update-TypeData -AppendPath $fullName
 		}
 	}
-	end
-	{
-	
-	}
 }