+++ PSFramework/internal/tepp/scripts/features.ps1
@@ -1,3 +1,3 @@
 ﻿Register-PSFTeppScriptblock -Name "PSFramework.Feature.Name" -ScriptBlock {
 	(Get-PSFFeature).Name
-}
+} -Global